+++ test.lua
@@ -1,19 +1,10 @@
 package.path="?.lua;?/init.lua;?.lua;?/?/init.lua;"..package.path
-multi,thread = require("multi"):init()
-local GLOBAL,THREAD = require("multi.integration.lanesManager"):init()
-func = THREAD:newFunction(function(test)
-    print(test)
-    THREAD.sleep(1)
-    return "Hello World!"
+multi = require("multi")
+GLOBAL, THREAD = require("multi.integration.lanesManager"):init()
+local stt = multi:newSystemThreadedTable("stt")
+stt["hello"] = "world"
+multi:newSystemThread("test thread",function()
+    local stt = GLOBAL["stt"]:init()
+    print(stt["hello"])
 end)
-func("Did it work").connect(function(...)
-    print(...)
-    --os.exit()
-end)
-local serv = multi:newService(function(self,data)
-    local name = thread.getRunningThread().Name
-    print(name)
-end)
-serv.Start()
-serv.SetPriority(multi.Priority_Low)
 multi:lightloop()